fix: index buffer

+++ Ignis/src/Ignis/Renderer/IndexBuffer.cpp
@@ -4,7 +4,7 @@
 
 namespace ignis
 {
-	std::shared_ptr<IndexBuffer> IndexBuffer::Create(const uint32_t* indices, size_t size)
+	std::shared_ptr<IndexBuffer> IndexBuffer::Create(const uint32_t* indices, uint32_t size)
 	{
 		switch (GraphicsAPI::GetType())
 		{
